DOTA-VLM/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0
opencv-python>=4.8.0
shapely>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
matplotlib>=3.7.0

# Optional - Object Detection
ultralytics>=8.0.0  # YOLO-OBB
mmdet>=3.0.0  # MMDetection (Oriented RCNN, etc.)
mmcv>=2.0.0

# Optional - Vision-Language Models
llava @ git+https://github.com/haotian-liu/LLaVA.git
accelerate>=0.24.0
bitsandbytes>=0.41.0
